refactor(be): seperate filter as a single file from exception themselves
