+++ backend/apps/admin/src/admin.module.ts
@@ -14,9 +14,9 @@ import {
 import { CacheConfigService } from '@libs/cache'
 import {
   BusinessExceptionFilter,
-  UnknownExceptionFilter,
-  apolloErrorFormatter
+  UnknownExceptionFilter
 } from '@libs/exception'
+import { apolloErrorFormatter } from '@libs/exception'
 import { pinoLoggerModuleOption } from '@libs/logger'
 import { PrismaModule } from '@libs/prisma'
 import { NoticeModule } from '@admin/notice/notice.module'
