+++ backend/libs/exception/src/business.exception.ts
@@ -1,20 +1,13 @@
 import {
-  Catch,
-  type ArgumentsHost,
-  type ExceptionFilter,
-  Logger
-} from '@nestjs/common'
-import {
   UnauthorizedException,
   type HttpException,
   NotFoundException,
   ConflictException,
   UnprocessableEntityException,
-  ForbiddenException,
-  InternalServerErrorException
+  ForbiddenException
 } from '@nestjs/common/exceptions'
 
-abstract class BusinessException extends Error {
+export abstract class BusinessException extends Error {
   name: string
 
   constructor(message: string) {
@@ -94,45 +87,5 @@ export class UnprocessableFileDataException extends UnprocessableDataException {
 export class ForbiddenAccessException extends BusinessException {
   convert2HTTPException(message?: string) {
     return new ForbiddenException(message ?? this.message)
-  }
-}
-
-@Catch(BusinessException)
-export class BusinessExceptionFilter implements ExceptionFilter {
-  private readonly logger = new Logger(BusinessExceptionFilter.name)
-
-  catch(exception: BusinessException, host: ArgumentsHost) {
-    const newException = exception.convert2HTTPException()
-    this.logger.log(newException)
-
-    const ctx = host.switchToHttp()
-    const res = ctx.getResponse()
-    const status = newException.getStatus()
-
-    res.status(status).json({
-      statusCode: status,
-      error: newException.name,
-      message: newException.message
-    })
-  }
-}
-
-@Catch()
-export class UnknownExceptionFilter implements ExceptionFilter {
-  private readonly logger = new Logger(UnknownExceptionFilter.name)
-
-  catch(exception: unknown, host: ArgumentsHost) {
-    const newException = new InternalServerErrorException()
-    this.logger.error(exception)
-
-    const ctx = host.switchToHttp()
-    const res = ctx.getResponse()
-    const status = newException.getStatus()
-
-    res.status(status).json({
-      statusCode: status,
-      error: newException.name,
-      message: newException.message
-    })
   }
 }
