+++ backend/libs/exception/src/index.ts
@@ -1,3 +1,4 @@
 export * from './business.exception'
 export * from './graphql-error-code'
 export * from './apollo-error-formatter'
+export * from './filter'
